requirements.txt
python-telegram-bot>=20.0
discogs-client
openpyxl
python-dotenv